+++ app/Http/Controllers/TagController.php
@@ -93,6 +93,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0){
+            session()->flash('warning',"This tag can not be deleted.");
+            return redirect()->back();
+        }
         $tag->delete($tag->id);
         return redirect(route('tags.index'));
     }
